Added a new layout- trigapless
Added a three-column layout. This is one of the only remaining things I
missed from i3.

Behavior:
The three column layout prefers to stick with three columns always. It
will allow two windows to have two colums, and lone windows still fill
the screen, but >3 window arrangements will /always/ have three columns.
Otherwise, it behaves like gaplessgrid, in that the "master" window will
vertically fill all available space, and only later windows will get
split if there's an odd number of columns.

The reason for this is that at ~8pt font on a 1080p or similar monitor,
three source files fit almost perfectly side-by-side. This allows for
really smooth single-screen multitasking. This layout does not allow for
further division, because at that point clients become squashed to some
degree and no longer are thirds of the screen, allowing for that 80col
layout. Up to this point, I had been using centeredmaster to accomplish
this, but it tended to favor the master slightly, and the stack windows
would have a few text columns cut off. Also, the stack worked in a way
that made sense if you really are focusing on the center, but since I'm
almost always multi-monitor, it took a lot of mental energy to rearrange
things as I like.

Eventually, the stack may end up in the middle of the screen for this
layout, and the top of the stack and master on the edges, but I'm not
decided if this would be a valuable addition.