Split project from DWM, updated man pages
Split project from DWM to allow someone to install headcannon-dwm
alongside dwm if desired, and also to indicate that hcdwm is no longer
tracking the DWM sources, as that is becoming increasingly impratical.

(finally) updated the man pages with correct information about the WM,
such as basic keyboard shortcuts, etc. This does not include an
exhaustive list of keys just yet, but it includes enough to get some use
out of the WM and does not misrepresent anything anymore.

Many thanks go to the DWM project- if you guys see this message, know
you have a fan in me!